Refactor to encode Level store using GraphQL scalar hints
This replaces levelup and encodingdown;
A minimal levelup layer is provided that ensures that everything
is promisified and errors in get are not thrown. Hence it currently
enforces only the used API.

Instead of encodingdown the src/encode helper is used in
ObjectTable to read field information and created minimal
serializer/deserializer functions.

All types had to be revamped and tests had to be changed to
match the new layer and serialization.